Refactor WebUiSettings into users/sub folder and add migration backfill
- Moves WebUiSettings entity to src/core/jupiter/core/users/sub/web_ui_settings/root.py
- Moves load/update use cases to users/sub/web_ui_settings/use_case/{load,update}.py
- Updates all imports to point to the new location
- Adds backfill INSERT in the migration so existing users get a default WebUiSettings row (use_night_mode=False)
